pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dmarc-msp"
version = "0.1.0"
description = "DMARC monitoring automation for Managed Service Providers"
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12,<1",
    "rich>=13,<14",
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.30,<1",
    "pydantic>=2,<3",
    "pydantic-settings>=2,<3",
    "sqlalchemy>=2,<3",
    "pyyaml>=6,<7",
    "httpx>=0.27,<1",
    "opensearch-py>=2,<3",
    "cloudflare>=3,<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23",
    "coverage>=7",
    "ruff>=0.5",
]
gcp = ["google-cloud-dns>=0.35"]
aws = ["boto3>=1.34"]
azure = ["azure-mgmt-dns>=8"]

[project.scripts]
dmarcmsp = "dmarc_msp.cli:app"

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
